Make RelationalReader consistent for non-English
RelationalReader's `get_file` strips away the language-specific directory prefix from DAT files when resolving them. This prefix is `Data/` for English and `Data/{language}/` for the rest.
The `[]` shortcut (`__get_item__`) unconditionally transforms `X.dat` and `Data/X.dat` into `Data/X.dat` which fails to strip anything in `get_file` when language is non-English.

This change alters the behaviour of the shortcut so that it transforms both forms of input into `Data/{language}/X.dat` so that the lookup will store the data file under the correct base filename.