Use pydds and Pillow for image conversion/slicing
With the porting of `process-image` from poe-utils to a Python extension at https://pypi.org/project/pydds/ we can now remove all subprocess invocations of both ImageMagick and `process-image`. This change still writes a number of intermediary files to disk and could be optimized to keep them in memory in some circumstances.

# Conflicts:
#	PyPoE/cli/exporter/wiki/parser.py
#	PyPoE/cli/exporter/wiki/parsers/item.py