Define stat translations and mod domain for templar relics
When a mod domain comes with a new stat translation text file it
needs to be explicitly mapped to reference that file.

Templar relics seem to use a previously unknown mod domain so we
also name that domain.

Gratuitous spec update for Mods included as some unused fields were
in the wrong order.