Restore Lua exporter for quest rewards
The data files for quest rewards have returned a few leagues ago, this
restores the old quest reward exporter with adaptations to the spec.

There's a new field which seems to indicate if a reward is for any
difficulty or specific to normal or hard mode (Ruthless).

Some of the past workarounds w.r.t item names might be possible to
remove but it's kept close to the original form for now.

Note that this does not include quest _vendor_ rewards, those seem to be
gone for good.