Add support for new hash algorithm
In Path of Exile 3.21.2 the hash algorithm for paths changed from
FNV1A64 to MurmurHash64A with a custom seed. The logic for hashing
pathnames also changed with the new paths always being lowercase,
without any trailing slashes and without the old suffix of `++`.

While the new seed is known in advance this code derives it
mathematically from the directory entries present in the index file to
make it more resilient against future seed changes.

This change introduces a new dependency on `mmhash2` to compute the new
hashes - existing installations need to re-run `pip install -e .`
to obtain and install the new dependency.