Skip path seed test and do insensitive indexing
The validation of computed hashes against the path reps is cute but has
some cost at startup for marginal value.
As the hash is unlikely to change in the future it should suffice to compute
the seed from the top entry as that is free.

Directory indexing is also made case-insensitive by lowercasing the
paths involved, inspired by @lvlvllvlvllvlvl's fork.