fix: Convert gem icons from/to sRGB when reshading
The source art for skill gem icons is encoded in 8-bit sRGB while the
shader source that the processing is inspired by works in a linear
colour space.

In order to process the art correctly, we need to convert from sRGB to
linear before running the shading algorithm and convert back from linear
to sRGB when we're done.

This change also lowers precision to 32-bit float when shading as 64-bit
floats are a bit over the top, 32-bit is also in-line with HLSL shading.