compose map icons like the game
With the introduction of five new purple T17 maps in 3.24 the old
approximation of how maps were composed from a base plate and a
colorized sigil was insufficient.

This change implements the relevant parts of the game UI shader for
tinting UI elements like the map items, ignoring parts that are not
used for these assets like saturation (always 1.0) and colour scaling
(also always 1.0). For reference, see `PShad` in `Draw2D.hlsl`.

The colour tint for the four different tier groups is obtained from the
game renderer for maximum correctness.

This also updates the defunct `atlas_icons` export to tint the sigils in
the same way.